Update skill tree mind map
We updated the USE skill set from the skill tree mind map according
to our axioms. This is an example of how we trimmed the tree and
selected skills. The links added to the nodes do not work currently.
The lead to the Markdown files which are still under construction.