Update Skill tree
This is a major update of the MindMap.
We pruned the tree a lot and worked on the skills directly.
Also, the markdown files are not correct anymore. We are still working
on them, but wanted to share the skill tree now so that further major
changes can be discussed. The Markdown files between the old map and the
updated one will be unified after this merge.
The wiki will be outdated for a while!