Update Skill tree (#23)
* Update Skill tree

This is a major update of the MindMap.
We pruned the tree a lot and worked on the skills directly.
Also, the markdown files are not correct anymore. We are still working
on them, but wanted to share the skill tree now so that further major
changes can be discussed. The Markdown files between the old map and the
updated one will be unified after this merge.
The wiki will be outdated for a while!

* Update USE tree

* Update BDA

* Update ADM tree

* Update PE

* Update SD

* Update K and SD again

* Update skill tree use 1.5

* Update Bug fixes in skill tree

* Update Bug fix

---------

Co-authored-by: Kevin Luedemann <kevin.luedemann@gwdg.de>